LayoutTests/canvas/philip/tests/2d.composite.clip.destination-atop-expected.txt

< [index] >
2d.composite.clip.destination-atop
fill() does not affect pixels outside the clip region.
References: 2d.composite.destination-atop
Actual output:
Expected output:

Passed
